Flip argument order of Witness::combine for and_v and and_b
Given an `and` instruction, miniscript produces a witness stack that
is wrongly ordered:

The left branch of an `and` instruction is evaluated first, which means
whatever witness stack is produced by this branch needs to come _after_
the witness stack of the right branch.

Unfortunately, Witness::combine simply concatenates the two vectors,
resulting in the witness stack for the left branch to end up _before_ the
one for the right branch.

This is what we currently have:

Witness in memory:

|----A-----|----B-----|--and(A,B)--|

Witness stack evaluation:

Round 0:
        Stack:          Instructions:

                        check(A) check(B)
        |----B-----|
        |----A-----|

Round 1:
        Stack:          Instructions:

                        |----B-----| check(A) check(B)
        |----A-----|

Round 2:
        Failure due to wrong signature.

With this patch, this changes to what we actually want:

Witness in memory:

|----B-----|----A-----|--and(A,B)--|

Witness stack evaluation:

Round 0:
        Stack:          Instructions:

                        check(A) check(B)
        |----A-----|
        |----B-----|

Round 1:
        Stack:          Instructions:

                        |----A-----| check(A) check(B)
        |----B-----|

Round 2:
        Stack:          Instructions:

                        check(B)
        |----B-----|

Round 3:
        Stack:          Instructions:

                        |----B-----| check(B)

Round 4:
        Script successfully executed!

Co-authored-by: Lucas Soriano del Pino <l.soriano.del.pino@gmail.com>