descriptor: add a method to get the scriptCode as per bip-0143
To generate valid Segwit sighashes, the `witness_script()` could be used
for P2WSH outputs's `scriptCode`. However a P2PKH Script would have to be
reconstructed by-hand for P2WPKH.

This unifies the way to get the `scriptCode` for any Segwit output
defined using a rust-miniscript descriptor.

Note that we don't bother checking for CODESEPARATOR as we don't produce
Scripts containing such opcode.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>