miniscript: make max_satisfaction_witness_elements() return an Option
For both Miniscript and Descriptor.

This -breaks the API and- leaves the choice of unwrapping to the caller,
as in max_dissatisfaction_witness_elements().

As we are going to check the number of witness items for standardness at
different stages we may call this on not yet sane Miniscript fragment,
which would have caused a panic.

This also corrects the Satisfaction for Terminal::False.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>