Remove inocrrect panic from satisfy
It is true that all thresh fragments must be `d`, but in
certain edge cases it is possible that we don't have the
information to dissatify it. For example, the preimage to
the PkH fragment. Although users should always use descriptors
and DescriptorPublicKey type when dealing with this and this
would not be an issue in practise we should still remove the panic