Workaround compiler goof in Policy::translate_pk
These previously untested methods result in a compiler error if you ever
tried to call them. This happens because if you put &mut in front of the
type when you recursively pass it to another call you are actually
creating a &mut &mut since the argument may already be a &mut. This ends
up in infinite &muts prefix. The compiler only figures this out when you call
the method (not at compile time of the library).

I fixed this by having an inner function take a &mut FnMut as the argument
and so breaking the chain of &mut. The public functions remains the same.