Cleanup context code with SigType enum
NoChecks context needs to know how to serialize the given public key or
where to lookup for it's signature (schnorr or ecdsa). The simplest way
to convey this information is to separate out NoChecks for ecdsa and
schnorr. When using NoChecksEcdsa the associated type
bitcoin::PublicKey or when using NoChecksSchnorr, the associated type
XOnlyPublicKey should take care of it.