Fix e/o bug in miniscript threshold correctness rules
This will not affect existing users of rust-miniscript as we were
previously more strict in the miniscript that we accepted. This
allows more more scripts to parsed as miniscripts.

This allows correctly tagging thresh z, o values as per the spec. For
example, miniscript with thresh(2,older(9),older(10)) to be tagged as `z`
and thresh(2,pk(),older()) as `o`

Found while doing a line to line comparison with c++ codebase and spec