Add DO_FEATURE_MATRIX section
In an effort to make CI more efficient add an if guard on env var
`DO_FEATURE_MATRIX` that:

- Runs tests for all features
- Runs tests for each feature individually
- Runs all the examples

Add default sanity tests (build and test) that will always be run
irrespective of env vars used when executing the script.

This is similar to how we test in `rust-secp256k1`.