Refactor implementations of TranslatePk
We recently refactored the `TranslatePk` traits but did not do _all_ the
implementations. We should be uniform in such things.

Refactor the implementations of `TranslatePk`. One non-uniformity
remains, use of generic `P` is almost universal except in the
`miniscript` module `Pk` is used. This patch leaves that as it is.