Use full tuple instead of KeySource
The `KeySource` type is a type alias from `bitcoin::bip32` for a
tuple (fingerprint, derivation_path). In other places in the code we use
the full tuple, we should be uniform and use one or the other. Elect to
use the tuple.

Use full tuple `(FingerPrint, DerivationPath)` instead of `KeySource`.