Use != 0 instead of == 1 for check of masked value
Recently [0] we added a check on a value masked against
`SEQUENCE_LOCKTIME_DISABLE_FLAG`, this check uses `x == 1` but x can
never be zero since `SEQUENCE_LOCKTIME_DISABLE_FLAG` is equal to
`0x80000000`.

Looking at the original commit it looks like we are trying to check for
a value _higher_ than the flag value, so we can mask then use `x != 0`
to see if the flag bit is enabled.

Found by clippy:

 error: incompatible bit mask: `_ & 2147483648` can never be equal to `1`

[0] commit: `commit a3dd9fe840c8cdae84e9d586cd9a9d6c195f0092`