Use the term 'abstract' instead of 'semantic'
Currently we are introducing a new term 'semantic' for what
is (according to the code comment) typically referred to as 'abstract'.
The justification give is that `abstract` is a reserved Rust keyword.

Rust provides a mechanism for using keywords as identifiers, one just
has to prefix with `r#`. Also, in this case we have aliases for abstract
and concrete `Policy` already so we barely ever need the `r#` trick.

Use the term 'abstract' instead of 'semantic', involves:

- Rename module `semantic` -> `abstract`
- Rename fuzz target `roundtrip_semantic.rs` -> `roundtrip_abstract.rs`
- Rename type `Semantic` -> `Abstract`
- Use `Abstract` everywhere instead of `semantic::Policy`