Do not write error if bitcoind process not found
We unconditionally kill `bitcoind` in the integration test script. This
is correct but when there is no `bitcoind` process running results in a
output `bitcoind: no process found`. This output is misleading because
it makes one think that `bitcoind` didn't start up correctly.

Send the stderr warning to `/dev/null`.