Shorten translatefpk identifier
Recently we refactored the translate pk traits, in doing so the
identifier fpk and fpkh were favoured over translatefpk, and
translatefpkh respectively. I missed a bunch of places.

Be uniform across the code base and use the identifiers fpk and fpkh.

This patch is the result of applying s/translatefpk/fpk/g to all source
files.

Refactor only, no logic changes.