Shorten path bitcoin::secp256k1 -> secp
Based off of an idea I had here [0] we can reduce the number of times
devs have to write 'secp256k1' since it is a tad clunky to write. Until
we get an alias in `rust-bitcoin` we can test run the usefulness of
such a thing here in `rust-miniscript` by using import aliases.

Shorten the use of `bitcoin::secp25k1::Foo` -> `secp::Foo` with no loss
of clarity.

The import alias can be later removed if/when `rust-bitcoin` adds an
alias as described in the discussion linked below.

- [0] https://github.com/rust-bitcoin/rust-bitcoin/discussions/984