Improve test coverage
In an effort to improve test coverage without being inefficient add a
new env variable `DO_FEATURE_MATRIX` to the test script. This is how we
do it over in `rust-secp256k1`, it enables good test coverage while not
having to run the whole matrix in every CI job.

Add cargo calls for each example because the current way we run the
examples does not work for those who configure the build output
directory to something other than the default.