Remove use of ref keyword
We currently use `ref` _a lot_. This patch falls victim to 'if you only
have a hammer everything looks like a nail' - I did a search and replace

  `s/ref //g` on every source file then fixed the build errors.

This almost certainly means I removed 'good' uses of `ref`.

An example of a code improvement is:

Before this patch is applied we had:

```rust
    pub fn into_pre_taproot_desc(self) -> Result<pretaproot::PreTaprootDescriptor<Pk>, Self> {
        match self {
            Descriptor::Bare(ref bare) => Ok(pretaproot::PreTaprootDescriptor::Bare(bare)),
            Descriptor::Pkh(ref pkh) => Ok(pretaproot::PreTaprootDescriptor::Pkh(pkh)),
            Descriptor::Wpkh(ref wpkh) => Ok(pretaproot::PreTaprootDescriptor::Wpkh(wpkh)),
            Descriptor::Sh(ref sh) => Ok(pretaproot::PreTaprootDescriptor::Sh(sh)),
            Descriptor::Wsh(ref wsh) => Ok(pretaproot::PreTaprootDescriptor::Wsh(wsh)),
            Descriptor::Tr(ref tr) => Err(Descriptor::Tr(tr)),
        }
    }
```

And afterwards we get:
```rust
    pub fn into_pre_taproot_desc(self) -> Result<pretaproot::PreTaprootDescriptor<Pk>, Self> {
        match self {
            Descriptor::Bare(bare) => Ok(pretaproot::PreTaprootDescriptor::Bare(bare)),
            Descriptor::Pkh(pkh) => Ok(pretaproot::PreTaprootDescriptor::Pkh(pkh)),
            Descriptor::Wpkh(wpkh) => Ok(pretaproot::PreTaprootDescriptor::Wpkh(wpkh)),
            Descriptor::Sh(sh) => Ok(pretaproot::PreTaprootDescriptor::Sh(sh)),
            Descriptor::Wsh(wsh) => Ok(pretaproot::PreTaprootDescriptor::Wsh(wsh)),
            Descriptor::Tr(tr) => Err(Descriptor::Tr(tr)),
        }
    }
```

A not so good result of this patch is having to write `*k` in a bunch of
places instead of `k`.

Pleas note patch includes such things as:
```rust
	WshInner::Ms(_ms) => DescriptorType::Wsh,
```

The `_ms` could become `_`, will do if this work is deemed valuable.