Remove script_pubkey from DescriptorTrait
In preparation for removing the `DescriptorTrait` move the
`script_pubkey` trait methods onto each individual descriptor. Replaces
the `spk` method. Note that once upon a time `spk` was an infallible
version of `script_pubkey` but now `script_pubkey` is itself infallible
so this patch just renames the `spk` method and removes the trait method.