Delete the DescriptorTrait
We never abstract over the `DescriptorTrait`, it is therefore
questionable for it to exist. We have recently removed a number of trait
methods out of the trait and implemented them directly on the respective
descriptors. We can now do the same with the remaining 'satisfaction'
trait methods.

Implement the various satisfaction methods on each descriptor as
required. Implement the methods directly on `Descriptor`. Delete
the `DescriptorTrait`.