WIP: Demonstrate the new timelock APIs
We are currently working on two new timelock APIs

- `LockTime`: https://github.com/rust-bitcoin/rust-bitcoin/pull/994
- `Sequence`: https://github.com/rust-bitcoin/rust-bitcoin/pull/1093

This PR demonstrates how we would use those APIs here in miniscript.

Includes improvements that are not yet in 1093. To test this out locally
you can grab the `WIP-sequence-and-lock-time` branch from my tree.
You'll have to have `rust-bitcoincore-rpc` locally and patch it to use
local `rust-bitcoin`.

This PR is not intended as a merge candidate.