Create configuration conditional bench
As we did in rust-bitcoin [0] create a configuration conditional bench
that we can use to guard bench mark code. This has the benefit of
making our features additive i.e., we can now test with --all-features
with a stable toolchain (currently this fails because of our use of the
test crate).

Please note, this patch maintains the current behaviour of turning on
the recovery and rand-std features when benching although I was
unable to ascertain why this is needed.

Maintains the use of `--features=compiler` in `test.sh` and in the
suggested incantation in `README.md`.

[0] - rust-bitcoin/rust-bitcoin#1092