Make DerivedDescriptorKey first-class citizen
Earlier we introduced DerivedDescriptorKey (an descriptor whose
wildcards are known to have been replaced with derivation indexes) to
eliminate wildcard derivation errors. This commit attempts to realise the
full benefits of this idea. I tried to keep backwards compatibility.

The naming of things no longer follows the logic of code so I'm going to
fix that in a follow up.