Rename "derive" things that are not doing derivation
The words "derive", "derived" and "derivable" were being used to
describe things that aren't stricly derivation. Derivation is the
process of taking cryptographic data and deterministically producing
something else given some inputs. This is how the term is used in bip32,
rust-bitcoin and more generally.

In rust-miniscript the term is also used to mean "replacing a wildcard
in a derivation path with a particular index". This can be quite
confusing. For example, `is_derivable` returns true when there is a
wildcard in the descriptor -- but it should return the opposite (given
the usual meaning)! You can't do a derivation because you don't know
the full derivation path yet (it's s got a wildcard in it). This commit
replaces all instances where "derive" is used in this way.