Use Key instead of MiniscriptKey
One of the main traits in this lib is `MiniscriptKey`, we can shorten
this to `Key` with no loss of meaning. This makes the whole codebase
more terse. Terse code, if clear, is easier to read because there is
less clutter. Also terseness assists the formatter and can reduce lines
of code.

This patch is the result of running

`s/MiniscriptKey/Key/g` on all source files and then running the
formatter.

To preserve backwards compatibility and make the library more
ergonomical to use; add a 'trait alias' of `MiniscriptKey` -> `Key`.