Use the new relative::LockTime module
Implement `check_older` using the new `relative::LockTime` API.

This code seems upside down, I'm not sure I have the naming correct,
i.e., I don't know which (`self` or arg) is doing the satisfying and
which is satisfied?

I believe the `<=` are going the correct direction, its the naming and
which sequence value to call the 'lock' that I'm struggling with.