Remove unstable feature
Currently it is not possible to run the test suite with `cargo test
--all-features`, this is because we use a feature called `unstable` to
enable the unstable `test` crate that is used for benchmarking.

There is another way to conditionally enable the test crate and guard
the benchmark code. We can use a custom configuration option `bench`
and then set it using `RUSTFLAGS='--cfg=bench`.