Introduce SPDX identifiers
Recently we introduced SPDX identifiers into `rust-bitcoin`. In brief,
SPDX identifiers are a way to fully describe the license of a source
file in a single line. In addition to this we use an optional "written
by" line to give attribution to the files original developer if they
wish to have it.

- Add links to the SPDX website in the readme
- Add SPDX line to files with no license blurb
- Shorten the author section to a single line while maintaining the
  authors (add Sanket's email address to one attribution line)
- Remove all the licence information in each file and replace it with an
  SPDX ID (see https://spdx.dev/ids/#how)