Use hash_newtype macro instead of the hash256 module
We currently implement the `hash256::Hash` type manually however there
is a macro in `bitcoin::hashes` that can do that for us.

For context, the `hash256::Hash` type exists solely to enable the type
to display forwards since `hashes::sha256d` displays backwards. The
`hash_newtype` macro takes an argument to configure display.