examples: Use SPDX licence identifiers
As we do in the main source directory; use SPDX license identifiers in
all of the examples files. The SPDX license is a balance between minimal
noise and having license coverage for those who want/like/need it.

Note, removes the "written by" section if it uses Andrews name because
he has stated elsewhere that this is ok. Leaves Thomas' attribution in
there.