examples: Clean up import statements
Examples are meant to showcase how to use `miniscript`; software using
`miniscript` as a dependency should not need to depend directly on
`bitcoin` since we re-export it from the `miniscript` crate root. Our
examples are more applicable if they do not assume a dependency on
`bitcoin` exists.

Use `use miniscript::bitcoin` to get at types/functions in `bitcoin`.