Remove fuzz from the workspace
I think that the fuzz crate is messing with the 1.41/1.47 builds. There
is no need for fuzzing to work with MSRV toolchain so instead of pinning
deps from fuzzing just remove the `fuzz` crate from the workspace.

Note chat `cargo check` from the crate root does not check the
`fuzz_target` source files like we hoped when adding `fuzz` to the
workspace.

I am not entirely sure why `cargo check` does not work but dependencies
from `fuzz` are still in the lock file.