iter: add module, copied in large from rust-simplicity
Introduces the new `iter` module which provides iteration abilities over
various tree-like structures. This is largely copied from
rust-simplicity, but is both simpler (because there is no sharing) and
more complicated (because we need to handle n-ary nodes, not just binary
ones).

This currently uses Arc<[T]> in a number of places; we should eventually
replace these with some sort of ArrayVec-type structure.