Rename traits MiniscriptKey and ScriptContext
In an effort to make the code more terse with no loss of clarity rename
the crate's main two traits `MiniscriptKey` to `Key` and `ScriptContext`
to `Context`.

The diff is big but can be verified by simply doing a search-and-replace
on `MiniscriptKey` -> `Key` and `ScriptContext` -> `Context`. Then run
`cargo +nightly fmt`. No other manual changes.