Remove written in clause
Recently we discussed in `rust-bitcoin` the pointlessness of the `\\
Written in ...` clause above the license terms. We came to the
conclusion that:

- It probably wasn't written then and who cares anyway because we patch
files all the time
- It almost certainly wasn't written by just this person because we have
a bunch of contributors
- Use of "the rust bitcoin devs" is just noise and no signal

Remove the "written in" clause if it contains "the rust bitcoin devs" or
Andrew's name (because he was part of the conversation in rust-bitcoin)

Do no remove clauses that have Sanket's name or Maxim's. Its up to them
if they want attribution so we leave it there for now.

Leave Andrew's name in `lib.rs` as a little nod-of-the-head to him as
the original author. I do not know the full relationship between Sanket
and Andrew and who did what so please forgive me Sanket if this is wrong.