Add ScriptContextError alias
Add an alias to the `miniscript::context` module to maintain backwards
comparability. Also, as for `ScriptContext` this may reduce potential
naming conflicts downstream because "context" is not that unique.

Note the alias is in `miniscript::context` but in `miniscript` there is
a re-export of `Contetxt` but not one of the error.