Fix stale variable names and comments
At some stage, it appears, we have changed from using a `HashMap` to a
`BTreeMap` but the variable names and comments still refer to hash map.
Note also that these comments are "what" comments (as opposed to "why"
comments) so they are not adding an value anyways.

Fix stale code by doing:

- Remove the "what" comments.
- Rename local variables from `hash_map` to plain old `map`