Remove PolicyArc
The `concrete::PolicyArc` type is identical to `Policy` except that it
uses `Arc` to wrap the nested `Policy` structs.

In preparation for implementing `TreeLike` and removing recursive
functions that process the `Policy` type remove the `PolicyArc` and add
`Arc` wrappers around the nested `Policy` fields.

Note, this patch does a bunch on cloning in the recursive functions.
This will dissapear when they are re-written to use tree iteration.

There are more optimizaitons available by taking `Arc<Policy<Pk>>` as a
parameter to functions in the `compiler` module, left for later.

The unit tests are bit klunky, much mapping of iterators to add `Arc`, I
think this will come out in the wash as we keep working.