Wrap `all_assets` for desc and add `count_assets` method
This commit wraps the all_assets function for all possible descriptor
types. It also adds `count_assets` function which can count the total
possible ways to obtain an asset for a given descriptor.

This API will help developer to count the possible ways. In-fact this
will also help in protecting that if some descriptor is very large and
has huge number of possible spend paths then it could simply be
dropped off and we can protect the network from traffic.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>