Add check_timelocks_within_range
Note that currently the `in_valid` function does a bunch of checks on
or/and/thresh that are also done in `from_tree_prob`.

Create a private function `check_timelocks_within_rang` that checks that
older/after locks are within valid ranges (above 0 and below 2^31).

Note this patch removes a bunch of logic but as can be seen by the
passing unit tests this code is still logically correct.