Rename semantic module to abstract
We used "semantic" so as not to clash with the reserved Rust keyword
"abstract", however there is a language feature to deal with such name
clashes - the raw identifier syntax `r#foo`.

Use `r#abstract` and rename the `policy/semantic.rs` file to
`policy/abstract.rs`. There is more re-naming to do to complete the
semantic -> abstract change but here we _only_ do the module name to
ease review.