Move abstract::Policy to crate/abstract/mod.rs
The abstract policy is different from the concrete policy because it is
only part of rust-miniscript and not miniscript in general. To help
differentiate it move it to a separate module.

Note to reviewers, `git` couldn't work out that this is basically a file
move so I've left the `src/policy/abstract.rs` file in the repo but it
is not included in the build. Reviewers can do `diff
src/policy/abstract.rs src/abstract/mod.rs` to ease review. It will be
removed in the next patch.