Order arms as in enum definition
Currently there are a bunch of places where when matching on the
`Policy` enum variants are matched in different order to the enum.
During recent work this was maintained to make diffs easier to review
however since we just modified a lot of this file lets clean it up while
we are here.

Move the match arms around so that they are ordered in the same order as
the variants are defined in the `Policy` enum.

Refactor only, no logic changes.