Add Arc to the semantic::Policy::Thresh vector
In preparation for removing recursive algorithms in the `semantic`
module add an `Arc` wrapper around the policies in the `Thresh` vector.

Note we use the more explicit `Arc::new` instead of `into` because a lot
of these should be removed when we remove recursion, later it may be
necessary to check that we are not creating new references when a clone
will suffice, the explicit `Arc::new` is easier to check. In tests
however just use `into`.