Implement write_descriptor macro
Currently we have a bunch of places in the code where we create a
checksum formatter wrapper then write a descriptor to it then write the
checksum. This can all be wrapped up in a macro, has to be a macro
because the number of args is variable.

This reduces the line count with no real loss of clarity
`write_descriptor!` is pretty self explanatory.