Add unit tests for normalized
The `semantic::Policy::normalized` function is recursive and has nested
processing of `Threshold`s. It is not trivial to fully grok.

The function is already tested elsewhere in the unit tests but we don't
have simple single purpose tests for it.

Add various unit tests for `normalized`, these serve as documentation of
the function as much as anything.