Remove default trait method implementations
Implementors of `MiniscriptKey` must reason about the three trait
methods, this implies the trait methods are required, not provided.

We are using the default impls to remove code duplication, this is an
abuse of the default impls. It makes the docs read incorrectly, by implying
that implementors _do not_ need to reason about these trait methods.

Remove default trait method implementations and manually implement the
trait methods for all current implementors.