Remove generics from Error by making fragment a String
This cause a reduction of IR lines emitted by the compiler of about 2%
(using cargo llvm-lines), so this should improve compile times.

I've seen using the parse example to check the produced binary size, but it's
a small snippet I am not sure it calls a representative portion of the API
surface.

From a library perspective this is equivalent since the field is used only
to be converted to string.

In theory externally someone could use the field in the error to do things so
it comes at a cost.