types: drop `Property` trait entirely
This trait is now only used on the `Type` type. Drop the trait, rename
the `from_` constructors, and remove unused parameters and type
parameters.

Most of the methods can now be constfns. Will do this in the next commit
since it's a bit annoying (mainly: can't use ?, have to be explicit
about early returns).