AbsLockTime: move to module, consolidate error checking
We have the type `AbsLockTime` whose purpose is to provide an ordering
for locktimes. This type is converted from strings and script values,
but we don't check validity during construction. Instead we have
multiple places where validity is manually checked. Fix this.

Note that this is an API-breaking change for people who are manually
constructing policies and Miniscripts since it removes a few ways to
unconditionally convert u32s into locktimes.