interpreter: rename "age" to "sequence"
The "age" from the transaction is actually just a sequence number.
Whether it is interpreted as an "age" depends on various other factors,
as described in BIP112.

In the next commits we will start moving much of our `Sequence` usage to
`relative::LockTime` and it will be helpful to have clearer names.