plan/satisfy: use relative::LockTime instead of Sequence to indicate relative locktimes
Eliminates some error logic. Also replaces manual implementations of
`is_implied_by` in a couple places.

This introduces a couple unwrap()s which will be removed in a later
commit, and which occur because there is a mismatch between the
`Sequence` type used in `satisfy::Satisfaction` and
`bitcoin::relative::LockTime` (and we can't switch this `Sequence`
because we depend on `Ord`, which will need a new newtype, which should
wait for another commit).

This *also* deletes a couple lines in the PSBT code which cites BIP 112
saying that if the sequence number passed to OP_CSV has the disable flag
set, to automatically pass `check_older`. With this code change, it is
no longer possible to produce such a situation, so we drop the check and
the citation.

I believe this is correct because no Miniscript-generated call to OP_CSV
should have the disable flag set. I believe the Miniscript spec is silent
on this and more-or-less says you can pass any crap you want as the
argument to `older`, but I am taking a stand here and think we should
Miniscripts with the disable flag set.