policy: use `RelLockTime` in semantic and concrete policies
This changes the signature of `at_age` to take a `relative::LockTime` to
be symmetric with `at_lock_time`. But because `relative::LockTime` has
no constructors this is kinda annoying to use.

See https://github.com/rust-bitcoin/rust-bitcoin/issues/2547

Anyway for now you can just construct a RelLockTime and tack .into()
onto it, which is fine as a workaround. We may want to update docs for
it.