compiler: eliminate unused error paths
Miniscript rules say that for andor(x,y,z), x must be dissatisfiable
(must have the 'd' property). This means that we cannot construct a
Miniscript with a non-dissatisfiable first child of an AndOr, because we
would get a type-checking error.

However we do this check again in CompilerExtData::and_or, for some
reason. This is actually the *only* non-threshold-related error path in
CompilerExtData::type_check, and getting rid of it will let us eliminate
a ton of errors. So remove this.

NEXT, we remove the two threshold-related error paths. These are the
checks in CompilerExtData::type_check_common which check for k == 0
and for k > n. I *think* these paths are possible, but they are not
tested, so we can remove them without breaking our tests. So we remove
them here, allowing us to eliminate a ton of Ok(), ? and unwraps. The
next commit will use the new Threshold type to make these error paths
actually impossible.