CI: Run fuzzer daily
Update the `fuzz/generate-files.sh` script to mimic that in
`rust-bitcoin` and generate a new fuzz job that runs the fuzz tests
daily for an hour.

Run the fuzz job the time that `rust-bitcoin` fuzz job finishes - I
don't know if that matters. I still don't know if there are org wide
limits or just repo limits but it doesn't hurt to run at a different
time.