Use rustfmt::skip instead of cfg_attr
Currently we use `cfg_attr` to disable the formatter so that the
attribute can be on an expression.

There was a new lint introduced into nightly just recently that checks
cfg args, this `cfg_attr` usage triggers it.

Use `rustfmt::skip` instead and put it on the function because it
doesn't currently work on expressions.