Depend on the new bitcoind-json-rpc group of crates
There is an effort to improve the state of affairs in regards to
integration testing extensively against multiple versions of Bitcoin
Core. As part of this do:

- Depend on the new `rust-bitcoind-json-rpc` crates
- Run the integration tests against most versions of Core since 0.17.1

(Note the latest supported version is currently 26.0)

This patch effects integration testing only and should hopefully help
with our upgrade process because I will personally make sure the new
crates are ready and tested during the rust-bitcoin RC cycle.