plan: make LoggerAssetProvider constructible
It looks like we have an example AssetProvider that is intended to be
used for testing/debugging. It looks useful but evidently has never been
used because it was impossible to construct.

Recent versions of rust nightly notice this and output a warning about
this, which is causing CI to fail, so we need to fix it.

My solution is to make the inner `Assets` public (and also a reference,
since there is no need to take ownership) so that people can easily
construct and deconstruct the type.