iter: get rid of Arc::clones and allocations for n-ary nodes
This somewhat complicates implementation of the trait, but not by much,
and eliminates a ton of refcount accesses and increments, and a ton of
Arc::clones.

As we will see, having an Arc<[T]> is somewhat restrictive in addition
to requiring allocations. Later we will want to replace this with an
enum that covers multiple kinds of arrays.

I was unsure whether to name the new associated type NAry or Nary. It
seems more natural to use NAry but the existing Tree::Nary variant uses
the lowercase spelling so I stuck with that. (And similarly used nary_
for the method names rather than n_ary.)