miniscript: non-recursive Clone
This is mostly just annoying/mechanical code. But it needs to be written
because the derive(Clone) logic, I think, was only doing a "shallow
clone" by cloning the Arcs in the Miniscript.

In my view this is undesirable behavior. If users want to do a shallow
clone, they can hold an Arc<Miniscript> and call Arc::clone on that. But
there are some cases where they might want to do a deep clone, and
currently there isn't really any way to do so (you can do it with the
translator API and an "identity" translator, but this is awkward to do
and will be slow because it redoes typechecking and rechecks for
duplicate keys).